requirements.txt
pytest>=7.0
pytest-html>=3.1.1
pytest-xdist>=3.0.0
requests>=2.28.0
pydantic>=1.10.7
allure-pytest